requirements.txt
Authlib==1.3.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
Flask-Assets==2.1.0
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
typing_extensions==4.12.0
urllib3==2.2.1
webassets==2.0
Werkzeug==3.0.3
WTForms==3.1.2
